+++ src/utility/wifi_drv.h
@@ -288,18 +288,18 @@ public:
 
     static int8_t downloadFile(const char* url, uint8_t url_len, const char *filename, uint8_t filename_len);
 
-    static int8_t fileOperation(uint8_t operation, const char *filename, uint8_t filename_len, size_t offset, uint8_t* buffer, size_t len);
+    static int8_t fileOperation(uint8_t operation, const char *filename, uint8_t filename_len, uint32_t offset, uint8_t* buffer, uint32_t len);
 
-    static int8_t readFile(const char *filename, uint8_t filename_len, size_t offset, uint8_t* buffer, size_t buffer_len) {
+    static int8_t readFile(const char *filename, uint8_t filename_len, uint32_t offset, uint8_t* buffer, uint32_t buffer_len) {
         return fileOperation(READ_FILE, filename, filename_len, offset, buffer, buffer_len);
     };
-    static int8_t writeFile(const char *filename, uint8_t filename_len, size_t offset, uint8_t* buffer, size_t buffer_len) {
+    static int8_t writeFile(const char *filename, uint8_t filename_len, uint32_t offset, uint8_t* buffer, uint32_t buffer_len) {
         return fileOperation(WRITE_FILE, filename, filename_len, offset, buffer, buffer_len);
     };
     static int8_t deleteFile(const char *filename, uint8_t filename_len)  {
         return fileOperation(DELETE_FILE, filename, filename_len, 0, NULL, 0);
     };
-    static int8_t existsFile(const char *filename, uint8_t filename_len, size_t* len)  {
+    static int8_t existsFile(const char *filename, uint8_t filename_len, uint32_t* len)  {
         int32_t length = 0;
         fileOperation(EXISTS_FILE, filename, filename_len, 0, (uint8_t*)&length, sizeof(length));
         *len = length;
